requirements.txt
future==0.15.2
requests==2.9.1
six==1.10.0
tld==0.7.6
PyYAML
lxml